test(ui): split catalog pagination fixtures and host cases
